backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "calorie-calculator-backend"
version = "1.0.0"
description = "基于GPT-4.1的食物识别和卡路里计算服务的后端部分"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Calorie Calculator Team" },
]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-multipart==0.0.6",
    "aiohttp==3.8.6",
    "python-dotenv==1.0.0",
    "pydantic==2.4.2",
    "gunicorn==21.2.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pytest",
    "pytest-asyncio",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88 